Update Trabalho Capítulo 7.py
Changhinho int variables to float

+++ python-program/Trabalho Capítulo 7.py
@@ -3,9 +3,9 @@
 
 # Capitulo 07 - Advanced concepts of algorithms and flowcharts for logic in decision making
 
-anos = int(input("Tempo como fumante (em anos)......:"))
-preco = int(input("Valor do maço....................:"))
-quant = int(input("Quantidade de maços por dia.......:"))
+anos = float(input("Tempo como fumante (em anos)......:"))
+preco = float(input("Valor do maço....................:"))
+quant = float(input("Quantidade de maços por dia.......:"))
 
 valor = ((anos * 365) * preco * quant)
 
@@ -15,4 +15,3 @@ elif valor > 20000 and valor <= 50000:
     print("Com o valor de {}, você poderia comprar uma bela moto".format(valor))
 else:
     print("Com o valor de {}, você poderia comprar um civic tunado".format(valor))
-
